WebApp/requirements.txt
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dnspython==2.1.0
dominate==2.6.0
eventlet==0.33.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-MQTT==1.1.1
Flask-SocketIO==5.1.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
paho-mqtt==1.5.1
psycopg2-binary==2.9.2
python-engineio==4.3.0
python-socketio==5.5.0
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
Flask-Script==2.0.6
Flask-Migrate==3.1.0              
Flask-SQLAlchemy==2.5.1              